+++ 103 + 309/Project/Source/SOC.c
@@ -115,14 +115,16 @@ void InitData_SOC(void)
 void App_SOC(void)
 {
 	static UINT32 s_u32LastAfeCurrentSampleSeq = 0U;
+	UINT32 u32AfeCurrentSeq;
 	UINT8 u8HasNewAfeSample;
 	UINT16 soc_ichg;
 	UINT16 soc_idsg;
 
-	u8HasNewAfeSample = (g_u32AfeCurrentSampleSeq != s_u32LastAfeCurrentSampleSeq) ? 1U : 0U;
+	u32AfeCurrentSeq = AfeCurrent_GetSeq();
+	u8HasNewAfeSample = (u32AfeCurrentSeq != s_u32LastAfeCurrentSampleSeq) ? 1U : 0U;
 	if (u8HasNewAfeSample)
 	{
-		s_u32LastAfeCurrentSampleSeq = g_u32AfeCurrentSampleSeq;
+		s_u32LastAfeCurrentSampleSeq = u32AfeCurrentSeq;
 		SOC_GetNetCurrentForCalc(g_stCellInfoReport.u16Ichg,
 								 g_stCellInfoReport.u16IDischg,
 								 &soc_ichg,